changes for show history

+++ src/main/java/SwitchAnalyzer/Commands/MOM/OnlyMOM/SaveSwitchCMD_MOM.java
@@ -8,7 +8,7 @@ import SwitchAnalyzer.miscellaneous.GlobalVariable;
 public class SaveSwitchCMD_MOM implements ICommand
 {
     public String switchName;
-    public int portNum;
+    public Long portNum;
     public void processCmd()
     {
         DBConnect.startRun(new DBSwitch(switchName, portNum));
